Moved Session model and schema to auth module
adjusted all files that needed Session model
